+++ src/test/java/lapr/project/model/ClientTest.java
@@ -69,16 +69,16 @@ class ClientTest {
 
     @Test
     void getnif() {
-        int nif = client.getnif();
-        int expResult = 123456789;
+        double nif = client.getnif();
+        double expResult = 123456789;
         assertEquals(expResult, nif);
     }
 
     @Test
     void setnif() {
         client.setnif(987654321);
-        int expResult = 987654321;
-        int nif = client.getnif();
+        double expResult = 987654321;
+        double nif = client.getnif();
         assertEquals(expResult, nif);
     }
 
@@ -130,16 +130,16 @@ class ClientTest {
 
     @Test
     void getCreditCardNumber() {
-        int cc = client.getCreditCardNumber();
-        int expResult = new BigInteger("1234567891057189").intValue();
+        long cc = client.getCreditCardNumber();
+        long expResult = new BigInteger("1234567891057189").longValue();
         assertEquals(expResult,cc);
     }
 
     @Test
     void setCreditCardNumber() {
-        client.setCreditCardNumber(new BigInteger("1829102918271622").intValue());
-        int cc = client.getCreditCardNumber();
-        int expResult = new BigInteger("1829102918271622").intValue();
+        client.setCreditCardNumber(new BigInteger("1829102918271622").longValue());
+        long cc = client.getCreditCardNumber();
+        long expResult = new BigInteger("1829102918271622").longValue();
         assertEquals(expResult,cc);
     }
